requirements.txt
python-telegram-bot
langchain
langchain-groq